requirements.txt
alembic==0.8.9
aniso8601==3.0.2
astroid==2.0.4
atomicwrites==1.1.5
attrs==18.1.0
bcrypt==3.1.2
certifi==2018.8.13
cffi==1.9.1
chardet==3.0.4
click==6.7
colorama==0.3.9
config==0.4.0
coverage==4.0.3
coverage-badge==0.2.0
docopt==0.6.2
flasgger==0.9.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.6
Flask-JWT==0.3.2
Flask-Migrate==2.0.2
Flask-RESTful==0.3.6
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
green==2.12.1
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
Mako==1.0.6
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
more-itertools==4.3.0
pluggy==0.7.1
psycopg2-binary==2.7.5
py==1.5.4
pycparser==2.17
PyJWT==1.4.2
pylint==2.1.1
pytest==3.7.1
pytest-cov==2.5.1
python-coveralls==2.9.1
python-editor==1.0.3
pytz==2018.5
PyYAML==3.13
rednose==1.3.0
requests==2.19.1
singleton==0.1.0
six==1.11.0
SQLAlchemy==1.1.4
termstyle==0.1.11
typed-ast==1.1.0
Unidecode==1.0.22
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.10.11
